% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aggregate_stats.R
\name{agg_per_read}
\alias{agg_per_read}
\title{function to make aggregate statistics from per read modified probabilities from megalodon}
\usage{
agg_per_read(dt)
}
\arguments{
\item{dt}{data.table object}
}
\value{
aggregate statistics of per read modified bases
}
\description{
function to make aggregate statistics from per read modified probabilities from megalodon
}
