% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_promoters_encode1kb}
\alias{gene_promoters_encode1kb}
\title{gene promoters from encode databasae, defined as 1000 bp of transcription start site}
\format{
\subsection{\code{gene_promoters_encode1kb}}{

A GenomicRanges object
\describe{
\item{seqnames}{chromosome name}
\item{ranges}{genomic start and end position}
\item{strand}{strand of the gene}
\item{ensgene.version}{ensembl gene id with version}
\item{ensgene}{ensembl gene id}
\item{entrez}{entrez id}
\item{biotype}{gene biotype}
\item{description}{gene description}
\item{symbol}{gene name}
}
}
}
\source{
\url{https://www.encodeproject.org/files/ENCFF493CCB/@download/ENCFF493CCB.bed.gz}
}
\usage{
gene_promoters_encode1kb
}
\description{
gene promoters from encode databasae, defined as 1000 bp of transcription start site
}
\keyword{datasets}
