% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bismark_coverage.R
\name{make_bismark_coverage}
\alias{make_bismark_coverage}
\title{function to export megalodon aggregate mod bases to bismark like coverage file
chromosome start end methylation_percentage count_methylated count_unmethylated}
\usage{
make_bismark_coverage(dt, file_stem = NULL)
}
\arguments{
\item{dt}{data.table object}

\item{file_stem}{stem of file name to write to}
}
\value{
aggregate statistics of 5mc and 5hmc from \verb{aggregate per read modified bases from megalodon};
}
\description{
function to export megalodon aggregate mod bases to bismark like coverage file
chromosome start end methylation_percentage count_methylated count_unmethylated
}
