% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylONT-package.R
\docType{package}
\name{methylONT-package}
\alias{methylONT}
\alias{methylONT-package}
\title{methylONT: Methlylation Analysis for Long Read DNA Sequencing}
\description{
Methylation analysis utils.
}
\author{
\strong{Maintainer}: Samuel Ahuno \email{ahunos@mskcc.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
