% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_megalodon_aggregate_mod_bases.R
\name{read_megalodon_aggregate_mod_bases}
\alias{read_megalodon_aggregate_mod_bases}
\title{function to make aggregate statistics from per read modified probabilities from megalodon}
\usage{
read_megalodon_aggregate_mod_bases(
  paths_aggregate_mod_bases,
  methyl_percentage_threshold,
  methyl_numb_reads_threshold
)
}
\arguments{
\item{paths_aggregate_mod_bases}{vector of paths to megalodon aggregate mod bases}

\item{methyl_percentage_threshold}{minimum percentage_methylation either 5mc or 5hmC to be called methylated otherwise unmethylated}

\item{methyl_numb_reads_threshold}{minimum number of reads to be called methylated otherwise unmethylated}
}
\value{
data.table object representing coverage of aggregate 5mc and 5hmc from megalodon, methylation percentage & state (5hmc, 5mc, U)
}
\description{
function to make aggregate statistics from per read modified probabilities from megalodon
}
\examples{
#read_megalodon_aggregate_mod_bases(paths_aggregate_mod_bases=c("dir/5hmc.bed", "dir/5mc.bed"), methyl_percentage_threshold=0.5, methyl_numb_reads_threshold=5)
}
